requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
boto3==1.14.25
botocore==1.17.25
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ConfigArgParse==1.2.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
entrypoints==0.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.6.2
geventhttpclient==1.4.4
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
greenlet==0.4.16
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
locust==1.1.1
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
xmltodict==0.12.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
